Create wmtk::TetOptimizerMesh by moving tetwild's implementation
The 3D half of the shared-base work, mirroring 9c79e153bd for 2D. Seeded by
MOVING TetWildMesh's code, so tetwild is byte-identical by construction and any
output movement is confined to the later commit where simwild-3D adopts the base.

Moved: VertexAttributes and FaceAttributes, the vertex and face attribute
collections, the envelope/solver/rounding/force-split members, MAX_ENERGY, and 21
methods -- is_inverted x2, is_inverted_f, get_quality x2, round,
round_all_vertices, get_max_avg_energy, is_edge_on_{surface,bbox},
face_is_on_surface, vertex_is_on_surface, edge_incident_surface_face_count,
get_order_of_vertex, get_faces_by_condition, output_faces,
gradation_smooth_sizing, active_quality_threshold, active_vertices,
swap_edge_{44,56}_energy, invariants, smooth_before, the vertex partitioners.

Cell attributes deliberately stay per-application: tetwild's tets carry winding
numbers and a flood-fill id, simwild's a CellTag set, and an AttributeCollection
member cannot change type in a derived class. The base reaches the one field it
shares through cell_quality()/set_cell_quality(). The plan called for measuring
that virtual on octocat before committing to it; no measurement is needed,
because of everything moved here only active_vertices() and get_max_avg_energy()
touch a cell attribute at all, and both are once-per-pass sweeps.

smooth_after does NOT move, though it is byte-identical between the two apps. Its
body is the shared optimization::smooth_vertex_3d template, which reaches
m_tet_attribute and smoothing_energy_envelope -- both per-application. A
call-graph check over the function bodies cannot see that dependency, because it
lives in the template; only the compiler catches it. smoothing_containment_envelope
did move, but is now virtual: tetwild's pull target and containment test coincide
in its single surface envelope, while simwild has a separate working envelope.

The base holds OptimizerParameters&; TetWildMesh adds Parameters& m_tet_params for
the 32 references to tetwild-only fields.

Gate: ctest 107/107. All 53 registered configs byte-identical against the run
after the previous commit. crown and octocat run 10 threads and are
nondeterministic there, so they were compared separately at num_threads 0 against
a build of the pre-refactor commit -- byte-identical.
