+++ components/tetwild/wmtk/components/tetwild/EdgeSwapping.cpp
@@ -92,7 +92,7 @@ size_t TetWildMesh::swap_all_edges_32()
     logger().info("edge swap prepare time: {:.4}s", time);
     size_t total_success = 0;
     SurfaceTopoSignature sig_before;
-    if (m_params.check_surface_topology) {
+    if (m_tet_params.check_surface_topology) {
         sig_before = surface_topology_signature();
     }
     auto setup_and_execute = [&](auto& executor) {
@@ -117,7 +117,7 @@ size_t TetWildMesh::swap_all_edges_32()
         time = timer.getElapsedTime();
         logger().info("edge swap operation time serial: {:.4}s", time);
     }
-    if (m_params.check_surface_topology) {
+    if (m_tet_params.check_surface_topology) {
         warn_if_surface_topology_changed(sig_before, "swap_all_edges_32");
     }
     return total_success;
@@ -147,7 +147,7 @@ bool TetWildMesh::swap_edge_before(const Tuple& t)
     // mistaken for interior because of stale m_is_on_surface flags (which would tear the surface).
     const int n_surf_faces = edge_incident_surface_face_count(t);
     if (n_surf_faces > 0) {
-        if (!m_params.allow_surface_swap) return false;
+        if (!m_tet_params.allow_surface_swap) return false;
         if (!prepare_surface_flip(t, incident_tets)) return false;
     }
 
@@ -482,7 +482,7 @@ size_t TetWildMesh::swap_all_edges_all()
     logger().info("edge swap prepare time: {:.4}s", time);
     size_t total_success = 0;
     SurfaceTopoSignature sig_before;
-    if (m_params.check_surface_topology) {
+    if (m_tet_params.check_surface_topology) {
         sig_before = surface_topology_signature();
     }
     auto setup_and_execute = [&](auto& executor) {
@@ -525,7 +525,7 @@ size_t TetWildMesh::swap_all_edges_all()
         time = timer.getElapsedTime();
         logger().info("edge swap operation time serial: {:.4}s", time);
     }
-    if (m_params.check_surface_topology) {
+    if (m_tet_params.check_surface_topology) {
         warn_if_surface_topology_changed(sig_before, "swap_all_edges_all");
     }
     return total_success;
@@ -545,7 +545,7 @@ size_t TetWildMesh::swap_all_edges_44()
     logger().info("edge swap 44 prepare time: {:.4}s", time);
     size_t total_success = 0;
     SurfaceTopoSignature sig_before;
-    if (m_params.check_surface_topology) sig_before = surface_topology_signature();
+    if (m_tet_params.check_surface_topology) sig_before = surface_topology_signature();
     auto setup_and_execute = [&](auto& executor) {
         executor.renew_neighbor_tuples = wmtk::renewal_edges;
         executor.priority = [&](auto& m, auto op, auto& t) { return m.get_length2(t); };
@@ -568,7 +568,7 @@ size_t TetWildMesh::swap_all_edges_44()
         time = timer.getElapsedTime();
         logger().info("edge swap 44 operation time serial: {:.4}s", time);
     }
-    if (m_params.check_surface_topology)
+    if (m_tet_params.check_surface_topology)
         warn_if_surface_topology_changed(sig_before, "swap_all_edges_44");
     return total_success;
 }
@@ -597,7 +597,7 @@ bool TetWildMesh::swap_edge_44_before(const Tuple& t)
     // direct incident-surface-face count so a genuine surface edge is never mistaken for interior.
     const int n_surf_faces = edge_incident_surface_face_count(t);
     if (n_surf_faces > 0) {
-        if (!m_params.allow_surface_swap) return false;
+        if (!m_tet_params.allow_surface_swap) return false;
         if (!prepare_surface_flip(t, incident_tets)) return false;
     }
 
@@ -610,21 +610,6 @@ bool TetWildMesh::swap_edge_44_before(const Tuple& t)
     face_attribute_tracker(*this, incident_tets, m_face_attribute, cache.changed_faces);
 
     return true;
-}
-
-double TetWildMesh::swap_edge_44_energy(
-    const std::vector<std::array<size_t, 4>>& tets,
-    const int op_case)
-{
-    double max_energy = -1;
-    for (const auto& vids : tets) {
-        if (is_inverted(vids)) {
-            return std::numeric_limits<double>::max();
-        }
-        const double e = get_quality(vids);
-        max_energy = std::max(max_energy, e);
-    }
-    return max_energy;
 }
 
 bool TetWildMesh::swap_edge_44_after(const Tuple& t)
@@ -694,7 +679,7 @@ size_t TetWildMesh::swap_all_edges_56()
     logger().info("edge swap 56 prepare time: {:.4}s", time);
     size_t total_success = 0;
     SurfaceTopoSignature sig_before;
-    if (m_params.check_surface_topology) sig_before = surface_topology_signature();
+    if (m_tet_params.check_surface_topology) sig_before = surface_topology_signature();
     auto setup_and_execute = [&](auto& executor) {
         executor.renew_neighbor_tuples = wmtk::renewal_edges;
         executor.priority = [&](auto& m, auto op, auto& t) { return m.get_length2(t); };
@@ -717,7 +702,7 @@ size_t TetWildMesh::swap_all_edges_56()
         time = timer.getElapsedTime();
         logger().info("edge swap 56 operation time serial: {:.4}s", time);
     }
-    if (m_params.check_surface_topology)
+    if (m_tet_params.check_surface_topology)
         warn_if_surface_topology_changed(sig_before, "swap_all_edges_56");
     return total_success;
 }
@@ -745,7 +730,7 @@ bool TetWildMesh::swap_edge_56_before(const Tuple& t)
     // incident-surface-face count so a genuine surface edge is never mistaken for interior.
     const int n_surf_faces = edge_incident_surface_face_count(t);
     if (n_surf_faces > 0) {
-        if (!m_params.allow_surface_swap) return false;
+        if (!m_tet_params.allow_surface_swap) return false;
         if (!prepare_surface_flip(t, incident_tets)) return false;
     }
 
@@ -758,21 +743,6 @@ bool TetWildMesh::swap_edge_56_before(const Tuple& t)
     face_attribute_tracker(*this, incident_tets, m_face_attribute, cache.changed_faces);
 
     return true;
-}
-
-double TetWildMesh::swap_edge_56_energy(
-    const std::vector<std::array<size_t, 4>>& tets,
-    const int op_case)
-{
-    double max_energy = -1;
-    for (const auto& vids : tets) {
-        if (is_inverted(vids)) {
-            return std::numeric_limits<double>::max();
-        }
-        const double e = get_quality(vids);
-        max_energy = std::max(max_energy, e);
-    }
-    return max_energy;
 }
 
 bool TetWildMesh::swap_edge_56_after(const Tuple& t)
